Разделение на классы
Спринт 5. Тема 2/4. Урок 4/7
