+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Artem Adiev on 06.12.2022.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
