+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Artem Adiev on 06.12.2022.
+//
+
+import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
