+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  QuestionFactoryProtocol.swift
+//  MovieQuiz
+//
+//  Created by Artem Adiev on 06.12.2022.
+//
+
+import Foundation
+
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion() -> QuizQuestion?
+}
